pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "human-in-the-loop-release-notes"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "openai",
  "pydantic",
  "python-dotenv",
  "mkdocs",
  "mkdocs-material",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
